fix: Handle returns in open ended questions
When generating bullet points, any newline we write is treated as a new
bullet point. This is problematic if the survey responses include a new
line.

To prevent this issue, we replace newlines in answers with \u000b which
is the line tabulation unicode character. In this way, we preserve the
breaks without creating new bullet points.